==> utils/postprocessing/03_reformat2embedding.ipynb <==
# Format dataframe for constrained embedding task

import os
import glob
import pandas as pd
from rdkit.Chem import PandasTools

pdb = '7JTO'
method = 'difflinker'

gen_folder = 'data/generated'
df = pd.read_csv(os.path.join(gen_folder, f'{pdb}_sampled_{method}_valid.csv'))
df.head()

### Filter by 2D filters

df_fil  = df[(~df.PAINS) & df.ring_arom]
len(df), len(df_fil)

### Filter by SAScore

# # keep rows with SA score < 4
df_fil = df_fil[df_fil.sa_linker < 4]
len(df_fil)

### Deduplicate

df_dedupl = df_fil.drop_duplicates(subset=['protac_smiles'])
len(df_dedupl)

### Remove molecules with formal charges

def is_charged(mol):
    for a in mol.GetAtoms():
        charge = a.GetFormalCharge()
        if charge != 0:
            return True
    return False

PandasTools.AddMoleculeColumnToFrame(df_dedupl, smilesCol='protac_smiles', molCol='protac_mol')
df_dedupl['charged'] = df_dedupl.protac_mol.apply(is_charged)
df_dedupl = df_dedupl[~df_dedupl.charged]
len(df_dedupl)

### Remove unnecessary columns

df_dedupl.columns

df_dedupl.drop(columns=['protac_mol'], inplace=True)
df_dedupl.reset_index(drop=True, inplace=True)
df_dedupl.to_csv(os.path.join(gen_folder, f'{pdb}_sampled_{method}_valid_fil.csv'), index=False)
